// 25032026/sce-front-auditoria/sce-front-auditoria/src/app/model/reportes/filtro-asistencia-mm.ts
export interface FiltroAsistenciaMiembroMesa {
    esquema?:string;
    usuario: string;
    idProceso: number;
    idEleccion: number;
    idCentroComputo: number;
    ubigeo: string;
    proceso: string;
    eleccion: string;
    mesa: string;
    codigoCentroComputo?: string;
    acronimo:string;
}